fix(3124): a dead cross-generation merge producer must never fall back to the concat
roborev blocker. `stream_generations_for_read`'s readiness oneshot has exactly two
failure modes, and they were flattened into one `Error`: a REPORTED `KWayMerger::new`
failure, and the blocking task DYING before it could signal. The retained JoinHandle
was then dropped UNJOINED, losing the JoinError and the panic message.

The consumer (`SSTableManager::scan_stream`) read any such `Err` as a merger
CONSTRUCTION failure and fell back to `spawn_fanout_merge` — the non-reconciling
token-order concat. So a panic inside `KWayMerger::new` returned a FULL-LENGTH,
UNRECONCILED result set (duplicated overwritten rows, resurrected deleted rows) as
`Ok` with only a `tracing::warn!`: silently WRONG data, a notch worse than the silent
truncation #3124 is about.

- new `MergeStreamSetupError` (child module `generation_merge/merge_stream_setup.rs`)
  distinguishes the two AT THE TYPE LEVEL: only `Construction` is
  `fallback_eligible()`; `ProducerDied` propagates.
- the dead-producer arm now JOINS the retained handle and reports the real cause,
  worded like the five sibling DIED sites (the JoinError's Display carries the panic
  message).
- new fault site `ScanTaskSite::CrossGenerationMerge`, checkpointed inside the
  construction window so a test can kill this producer deterministically.
- #1116 campsite: this file was over the 800-line target, so the two responsibilities
  split out here (setup outcome + the read-visibility unit pins) bring it back under.
