+++ cqlite-core/src/storage/sstable/generation_merge.rs
@@ -37,8 +37,11 @@
 //! The whole module is gated on `write-support` at its `mod` declaration in the
 //! parent (`sstable/mod.rs`).
 //!
-//! NOTE (#1116 file-size): already over the 800-line target on `origin/main`; the
-//! #2063 admission acquires add a few lines. Splitting is tracked under epic #1116.
+//! NOTE (#1116 file-size): this file was over the 800-line target on `origin/main`.
+//! Touching it for issue #3124 split two responsibilities out into child modules —
+//! the streaming path's setup-outcome type (`merge_stream_setup`) and the
+//! read-visibility filter's unit pins (`read_shadow_tests`) — which brings it back
+//! under the target.
 
 use std::collections::{HashMap, HashSet};
 use std::path::PathBuf;
@@ -57,6 +60,11 @@ use super::{reader, scan_merge};
 use crate::storage::write_engine::merge::{CellData, KWayMerger, MergeEntry, MergeStep, RowData};
 use crate::types::{CellWriteMetadata, TableId as CqlTableId};
 use crate::{Result, RowCells, RowKey, ScanRow, Value};
+
+#[cfg(not(feature = "tombstones"))]
+mod merge_stream_setup;
+#[cfg(not(feature = "tombstones"))]
+pub(super) use merge_stream_setup::MergeStreamSetupError;
 
 /// One reconciled metadata row inside the merge task, before per-cell metadata is
 /// attached: `(partition key bytes, ScanRow row carrier, [(column,
@@ -570,25 +578,25 @@ fn push_metadata_rows(
 /// guardrail).
 ///
 /// Construction (`KWayMerger::new`, which opens the input files) happens on the
-/// blocking task; its success/failure is signalled back over a oneshot BEFORE any
-/// streaming, so the caller can FALL BACK to the lazy per-reader streaming merge on
-/// a construction error exactly as the materializing `scan` falls back to
-/// concatenation. A runtime `step()` error mid-stream is delivered as an `Err`
-/// item on the channel (the consumer sees it); a task that DIES instead of reporting
-/// is caught by the returned [`reader::RowScanStream`]'s join (issue #3124).
+/// blocking task and is signalled back over a oneshot BEFORE any streaming — and the
+/// TWO ways that can fail are kept apart by the returned
+/// [`MergeStreamSetupError`] rather than flattened into one `Error`. Only a
+/// REPORTED construction failure is `fallback_eligible`; a producer that DIED
+/// without signalling is joined here and reported as `ProducerDied`, which the caller
+/// must not answer with the non-reconciling concat (issue #3124, roborev — see that
+/// type's module doc for why the flattened version returned wrong data).
 ///
-/// This is a deliberate, documented error-path asymmetry (issue #1579): the
-/// caller's fallback-to-concatenation only ever applies to the CONSTRUCTION
-/// failure above (nothing has been streamed yet, so falling back cannot mix
-/// reconciled and unreconciled rows). A `step()` failure after some partitions
-/// were already emitted downstream is NEVER retried/fallen-back — it ends the
-/// stream via the `Err` channel item instead. That is safer than it sounds: the
-/// materializing `merge_generations_for_read` has no equivalent mid-collection
-/// failure signal — a `step()` error there simply propagates the whole call as
-/// `Err` before any partial `Vec` is returned to the caller. The streaming
-/// driver's `Err` item preserves the same "no half-reconciled result surfaces
-/// silently" guarantee while still letting the caller observe exactly how many
-/// good rows it already received.
+/// A `step()` error mid-stream is delivered as an `Err` item on the channel, and a
+/// task that dies mid-stream is caught by the returned [`reader::RowScanStream`]'s
+/// join (issues #3106/#3124).
+///
+/// Deliberate, documented error-path asymmetry (issue #1579): the caller's
+/// fallback-to-concatenation applies ONLY to the construction failure above (nothing
+/// has been streamed, so falling back cannot mix reconciled and unreconciled rows). A
+/// `step()` failure after partitions were already emitted is NEVER retried — the
+/// `Err` item ends the stream, giving the caller an honest cutoff instead of a
+/// silently half-reconciled table (the materializing `merge_generations_for_read` has
+/// no equivalent mid-collection signal; it propagates the whole call as `Err`).
 #[cfg(not(feature = "tombstones"))]
 pub(super) async fn stream_generations_for_read(
     reader_list: &[Arc<reader::SSTableReader>],
@@ -596,16 +604,27 @@ pub(super) async fn stream_generations_for_read(
     start_key: Option<&RowKey>,
     end_key: Option<&RowKey>,
     buffer_size: usize,
-) -> Result<reader::RowScanStream> {
+) -> std::result::Result<reader::RowScanStream, MergeStreamSetupError> {
     let start_key = start_key.cloned();
     let end_key = end_key.cloned();
     let paths = ordered_generation_paths(reader_list);
     let schema = schema.clone();
 
+    // Issue #3124 (site 5): this task's ONE test-only fault checkpoint needs the
+    // reader identity INSIDE the task, so capture it as an owned scope (a zero-sized
+    // no-op in a production build). Any generation's path identifies the table
+    // directory a test scopes to.
+    let fault_scope = crate::storage::producer_fault::FaultScope::capture(|| {
+        paths.first().cloned().unwrap_or_default()
+    });
+
     let (ready_tx, ready_rx) = oneshot::channel::<Result<()>>();
     let (out_tx, out_rx) = mpsc::channel::<Result<(RowKey, ScanRow)>>(buffer_size.max(1));
 
     let task = tokio::task::spawn_blocking(move || {
+        // Inside the construction window, i.e. BEFORE either readiness arm runs, so
+        // an armed fault reproduces exactly "the producer died without signalling".
+        fault_scope.checkpoint(crate::storage::producer_fault::ScanTaskSite::CrossGenerationMerge);
         // Issue #1849: capture the read-time TTL clock ONCE per scan.
         let shadow = ReadShadow::new(&schema, now_epoch_secs());
         let mut merger = match KWayMerger::new(paths, &schema) {
@@ -663,9 +682,14 @@ pub(super) async fn stream_generations_for_read(
 
     match ready_rx.await {
         Ok(Ok(())) => Ok(reader::RowScanStream::new(out_rx, task)),
-        Ok(Err(e)) => Err(e),
-        Err(_) => Err(crate::Error::Storage(
-            "cross-generation streaming merge task ended before signalling readiness".to_string(),
+        Ok(Err(e)) => Err(MergeStreamSetupError::Construction(e)),
+        // `ready_tx` is dropped-WITHOUT-send on exactly one condition: the blocking
+        // task unwound before either readiness arm ran. So this `Err` ⟺ a dead
+        // producer — JOIN the retained handle to recover the real cause (the panic
+        // message) instead of dropping it unjoined, and report it as the
+        // fallback-INELIGIBLE variant (issue #3124, roborev).
+        Err(_) => Err(MergeStreamSetupError::ProducerDied(
+            merge_stream_setup::dead_merge_producer_error(task.await),
         )),
     }
 }
@@ -680,152 +704,4 @@ fn ordered_generation_paths(reader_list: &[Arc<reader::SSTableReader>]) -> Vec<P
 }
 
 #[cfg(test)]
-mod tests {
-    //! Issue #1849: deterministic pins for the post-merge read-visibility filter,
-    //! independent of on-disk fixtures. The end-to-end multi-generation proof lives
-    //! in `tests/issue_1849_multigen_tombstone_ttl_shadow.rs`.
-    use super::*;
-
-    fn shadow(now_secs: i64) -> ReadShadow {
-        ReadShadow {
-            now_secs,
-            key_columns: HashSet::new(),
-        }
-    }
-
-    fn shadow_with_keys(now_secs: i64, keys: &[&str]) -> ReadShadow {
-        ReadShadow {
-            now_secs,
-            key_columns: keys.iter().map(|k| k.to_string()).collect(),
-        }
-    }
-
-    fn live_cell(column: &str, ts: i64) -> CellData {
-        CellData::new(column.to_string(), Value::Integer(1), ts)
-    }
-
-    fn expiring_cell(column: &str, ts: i64, ldt: i32) -> CellData {
-        let mut c = CellData::new(column.to_string(), Value::Integer(1), ts);
-        c.ttl = Some(60);
-        c.local_deletion_time = Some(ldt);
-        c
-    }
-
-    /// A TTL-expired data cell (past `localDeletionTime`) is dropped while a
-    /// live-forever sibling survives; the row itself stays (not partition-shadowed).
-    #[test]
-    fn filter_live_drops_ttl_expired_cell_keeps_live() {
-        let now = 2_000_000i64;
-        let cells = vec![
-            live_cell("name", 100),
-            expiring_cell("token", 100, 1_000), // expired: 1000 <= now
-        ];
-        let kept = shadow(now).filter_live(None, cells).expect("row visible");
-        let names: Vec<&str> = kept.iter().map(|c| c.column.as_str()).collect();
-        assert_eq!(names, vec!["name"], "expired `token` must be dropped");
-    }
-
-    /// A cell tombstone is never live data (dropped like the single-gen path).
-    #[test]
-    fn filter_live_drops_cell_tombstone() {
-        let mut tomb = CellData::new("gone".to_string(), Value::Integer(0), 100);
-        tomb.value = Value::Tombstone(Box::new(crate::types::TombstoneInfo {
-            deletion_time: 100,
-            tombstone_type: crate::types::TombstoneType::CellTombstone,
-            local_deletion_time: 0,
-            ttl: None,
-            range_start: None,
-            range_end: None,
-        }));
-        let kept = shadow(0)
-            .filter_live(None, vec![live_cell("keep", 100), tomb])
-            .expect("row visible");
-        assert_eq!(kept.len(), 1);
-        assert_eq!(kept[0].column, "keep");
-    }
-
-    /// A row whose every data cell is shadowed by the partition tombstone (all data
-    /// older than `markedForDeleteAt`) is hidden entirely (`None`); a row with a cell
-    /// strictly newer than the deletion survives.
-    #[test]
-    fn filter_live_partition_cover_hides_fully_shadowed_row() {
-        let cover = Some(2_000i64);
-        // All data older/equal to the cover → whole row hidden.
-        let hidden =
-            shadow(0).filter_live(cover, vec![live_cell("a", 1_000), live_cell("b", 2_000)]);
-        assert!(hidden.is_none(), "fully-shadowed row must be hidden");
-        // A cell strictly newer than the cover → row survives (and keeps newer cell).
-        let kept = shadow(0)
-            .filter_live(cover, vec![live_cell("a", 1_000), live_cell("b", 3_000)])
-            .expect("row visible");
-        let names: Vec<&str> = kept.iter().map(|c| c.column.as_str()).collect();
-        assert_eq!(names, vec!["b"], "older `a` shadowed, newer `b` survives");
-    }
-
-    /// A row kept alive by a DATA cell newer than the partition tombstone (a
-    /// post-delete resurrecting UPDATE) must retain its clustering-key pseudo-cell
-    /// even when that pseudo-cell's own write timestamp is `<= cover`. The shadow/
-    /// expiry drop is STRUCTURALLY skipped for key columns (matching the single-gen
-    /// path + the merger), so the surviving row is never emitted missing its
-    /// clustering-key value (issue #1849; roborev multi-gen finding).
-    #[test]
-    fn filter_live_keeps_clustering_key_under_partition_cover() {
-        let cover = Some(2_000i64);
-        // `ck` is a clustering pseudo-cell with ts <= cover (would be dropped if the
-        // shadow test were applied to it); `data` is a newer resurrecting cell; `old`
-        // is a stale data cell that MUST be shadowed away.
-        let cells = vec![
-            live_cell("ck", 1_000),
-            live_cell("old", 1_000),
-            live_cell("data", 3_000),
-        ];
-        let kept = shadow_with_keys(0, &["ck"])
-            .filter_live(cover, cells)
-            .expect("row survives via newer `data` cell");
-        let names: Vec<&str> = kept.iter().map(|c| c.column.as_str()).collect();
-        assert!(
-            names.contains(&"ck"),
-            "clustering-key pseudo-cell must be retained, got {names:?}"
-        );
-        assert!(
-            names.contains(&"data"),
-            "newer resurrecting data cell must survive, got {names:?}"
-        );
-        assert!(
-            !names.contains(&"old"),
-            "stale data cell shadowed by the partition tombstone must be dropped, got {names:?}"
-        );
-    }
-
-    /// A clustering-key pseudo-cell does NOT by itself keep a row alive: a row whose
-    /// only non-key data is fully shadowed is still hidden even though the key cell is
-    /// retained-eligible (the key cell is excluded from the `max_data_ts` fold).
-    #[test]
-    fn filter_live_key_cell_does_not_resurrect_shadowed_row() {
-        let cover = Some(2_000i64);
-        let cells = vec![live_cell("ck", 1_000), live_cell("old", 1_000)];
-        let hidden = shadow_with_keys(0, &["ck"]).filter_live(cover, cells);
-        assert!(
-            hidden.is_none(),
-            "a key pseudo-cell alone must not keep a fully-shadowed row visible"
-        );
-    }
-
-    /// AC4 post-2038: an expiring cell whose `localDeletionTime` is a post-2038
-    /// instant stored as a NEGATIVE `i32` bit pattern is reinterpreted UNSIGNED, so
-    /// it reads as a large FUTURE expiry (not wrongly wrapped negative / long-expired).
-    #[test]
-    fn cell_expiry_secs_reinterprets_post_2038_unsigned() {
-        // 2039-ish epoch seconds > i32::MAX, stored as the wrapping `as i32`.
-        let future: i64 = 2_200_000_000;
-        let stored = future as u32 as i32; // negative bit pattern
-        let c = expiring_cell("token", 100, stored);
-        assert_eq!(
-            cell_expiry_secs(&c),
-            Some(future),
-            "post-2038 LDT must reinterpret unsigned to a future expiry"
-        );
-        // A non-expiring (no-TTL) cell has no expiry.
-        assert_eq!(cell_expiry_secs(&live_cell("x", 100)), None);
-    }
-}
+mod read_shadow_tests;
